Restructured, no more redux for now
