+++ src/app/components/Composer/components/ContentEditor/index.js
@@ -1,0 +1,46 @@
+// tools
+import React from "react"
+import PropTypes from "prop-types"
+import { Editor, Raw } from "slate"
+
+
+
+import { plugins } from "./plugins"
+import { schema } from "./schema"
+import { loadContent } from "../../helpers/loader"
+import { saveContent } from "../../helpers/saver"
+
+
+// return
+export class ContentEditor extends React.Component {
+	state = {
+		state: Raw.deserialize(loadContent(), {terse: true}),
+		schema
+	}
+  onChange = state => this.setState({ state })
+  onDocumentChange = saveContent
+	render() {
+		return (
+			<Editor
+				plugins={						plugins }
+				schema={						this.state.schema }
+				state={							this.state.state }
+				onChange={					this.onChange }
+				onPaste={						this.onPaste }
+				onDocumentChange={	this.onDocumentChange }
+				onKeyDown={					this.onKeyDown }
+			/>
+		)
+	}
+}
+
+// declare
+Editor.propTypes = {
+	plugins:					PropTypes.array,
+	schema:						PropTypes.object,
+	state: 						PropTypes.object,
+	onChange:					PropTypes.func,
+	onPaste:					PropTypes.func,
+	onDocumentChange:	PropTypes.func,
+	onKeyDown:				PropTypes.func,
+}
