+++ src/app/components/Composer/components/ContentEditor/plugins/index.js
