+++ src/app/components/Composer/components/ContentEditor/plugins/linkify.js
