+++ src/app/components/Composer/components/ContentEditor/plugins/mark-hotkey.js
