+++ src/app/components/Composer/components/ContentEditor/plugins/paste-html.js
