+++ src/app/components/Composer/components/ContentEditor/rules.js
