+++ src/app/components/Composer/components/HeaderEditor/components/InputTitle/index.js
@@ -16,7 +16,7 @@ export class InputTitle extends React.Component {
   }
   handleChange(event) {
     this.setState({ value: toTitleCase(event.target.value) })
-    this.props.onChange(event.target.value)
+    this.props.onChange(event.target.value)    
   }
   render() {
     return <Textarea
