+++ src/app/components/Composer/components/HeaderEditor/state.json
@@ -1,0 +1,1 @@
+{ "title": "", "subtitle": "" }
