+++ src/app/views/Submit/Compose/index.js
@@ -1,0 +1,11 @@
+// tools
+import React from "react"
+
+// components
+import { Composer } from "../../../components/Composer"
+
+
+// render
+export const Compose = props => {
+	return( <Composer/> )
+}
